% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MG_RastPipeline.R
\name{ag.Class}
\alias{ag.Class}
\title{Extract class}
\usage{
ag.Class(list)
}
\arguments{
\item{list}{list of samples}
}
\description{
description
}
\examples{
ag.Class()
}
\keyword{download}
\keyword{mg-RAST}
\keyword{taxonomy}
