% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MG_RastPipeline.R
\name{ag.Domain}
\alias{ag.Domain}
\title{extract domain}
\usage{
ag.Domain(list)
}
\arguments{
\item{list}{list of samples}
}
\description{
extract domain level information from taxonomy mg-rast download
}
\examples{
ag.Domain()
}
\keyword{download}
\keyword{mg-RAST}
\keyword{taxonomy}
