% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MG_RastPipeline.R
\name{ag.family}
\alias{ag.family}
\title{Extract family}
\usage{
ag.family(list)
}
\arguments{
\item{list}{list of samples}
}
\description{
description
}
\examples{
ag.family()
}
\keyword{download}
\keyword{mg-RAST}
\keyword{taxonomy}
