% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MG_RastPipeline.R
\name{ag.Genus}
\alias{ag.Genus}
\title{extract genus}
\usage{
ag.Genus(list)
}
\arguments{
\item{list}{list of samples}
}
\description{
description
}
\examples{
ag.Genus()
}
\keyword{download}
\keyword{mg-RAST}
\keyword{taxonomy}
