% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MG_RastPipeline.R
\name{ag.Order}
\alias{ag.Order}
\title{Extract Order}
\usage{
ag.Order(list)
}
\arguments{
\item{list}{list of samples}
}
\description{
description
}
\examples{
ag.Order()
}
\keyword{download}
\keyword{mg-RAST}
\keyword{taxonomy}
