% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MG_RastPipeline.R
\name{ag.Phylum}
\alias{ag.Phylum}
\title{Extract phylum}
\usage{
ag.Phylum(list)
}
\arguments{
\item{list}{list of samples}
}
\description{
description
}
\examples{
ag.Phylum()
}
\keyword{download}
\keyword{mg-RAST}
\keyword{taxonomy}
