% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MG_RastPipeline.R
\name{ag.Species}
\alias{ag.Species}
\title{extract species}
\usage{
ag.Species(list)
}
\arguments{
\item{list}{list of samples}
}
\description{
description
}
\examples{
ag.Species()
}
\keyword{download}
\keyword{mg-RAST}
\keyword{taxonomy}
