% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MG_RastPipeline.R
\name{ag.tax}
\alias{ag.tax}
\title{aggregate taxonomy table}
\usage{
ag.tax(l.taxa, f)
}
\arguments{
\item{list}{list of samples}
}
\description{
aggregate mg-rast taxonomy data across samples into table at specified taxonomy rank
}
\examples{
ag.tax()
}
\keyword{download}
\keyword{mg-RAST}
\keyword{taxonomy}
