% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MG_RastPipeline.R
\name{condense}
\alias{condense}
\title{condense}
\usage{
condense(df, samkey, sam)
}
\arguments{
\item{df}{data frame of taxonomy or functional annotations}

\item{samkey}{sample key of unique mg-rast IDs in one column and sample IDs in another}

\item{sam}{sample metadata with rows being unique to each sample}
}
\description{
combine annotations based on sample ID
}
\examples{
condense()
}
\keyword{download}
\keyword{mg-RAST}
\keyword{taxonomy}
