% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MG_RastPipeline.R
\name{ConnStat2}
\alias{ConnStat2}
\title{construct table of connectivity values}
\usage{
ConnStat2(list, num)
}
\arguments{
\item{x}{phyloseq object}

\item{cat}{categorical variable to subset phyloseq object}
}
\description{
constructs network based explicitly on number of expected edges.
Challenge is in defining appropriate level of edges for interpretable topology.
Most nodes have at least one highly significant correlation to another; thus with very few edges, we get all the nodes, and as we decrease significance threshold, the network fills out.
}
\examples{
netstat()
}
\keyword{igraph}
\keyword{metric}
\keyword{summaries}
