% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MG_RastPipeline.R
\name{DFTax}
\alias{DFTax}
\title{combine tables and condense}
\usage{
DFTax(l)
}
\arguments{
\item{x}{list of mg-rast accession numbers}

\item{auth}{authentication code for mg-rast}

\item{ont}{level of ontology. acceptable: Subsystems, COG, KO, NOG}
}
\description{
download functional annotations of many files, combine into a table
}
\examples{
DFTax()
}
\keyword{download}
\keyword{function}
\keyword{mg-RAST}
