% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MG_RastPipeline.R
\name{download.F}
\alias{download.F}
\title{download function annotation}
\usage{
download.F(x, level, ont)
}
\arguments{
\item{x}{list of mg-rast accession numbers}

\item{ont}{level of ontology. acceptable: Subsystems, COG, KO, NOG}

\item{auth}{authentication code for mg-rast}
}
\description{
description
}
\examples{
download.F()
}
\keyword{download}
\keyword{function}
\keyword{mg-RAST}
