% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MG_RastPipeline.R
\name{download.O}
\alias{download.O}
\title{download taxonomic data}
\usage{
download.O(sample, auth)
}
\arguments{
\item{sample}{mg-rast ID, or list of IDs}

\item{auth}{mg-rast authentication code}
}
\description{
this functio wraps and summarizes the download taxa so that many samples can be processed together
}
\examples{
download.O()
}
\keyword{download}
\keyword{mg-rast}
\keyword{taxonomy}
