% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MG_RastPipeline.R
\name{downloadFunc}
\alias{downloadFunc}
\title{download functions}
\usage{
downloadFunc(x, level, ont)
}
\arguments{
\item{x}{list of mg-rast accession numbers}

\item{ont}{level of ontology. acceptable: Subsystems, COG, KO, NOG}

\item{auth}{authentication code for mg-rast}
}
\description{
download functional annotations of many files, combine into a table
}
\examples{
downloadFunc()
}
\keyword{download}
\keyword{function}
\keyword{mg-RAST}
