% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MG_RastPipeline.R
\name{downloadTaxa}
\alias{downloadTaxa}
\title{download taxonomic data}
\usage{
downloadTaxa(sample, auth)
}
\arguments{
\item{sample}{mg-rast ID}

\item{auth}{mg-rast authentication code}
}
\description{
download taxonoic data for a single annotation run
}
\examples{
downloadTaxa()
}
\keyword{download}
\keyword{mg-RAST}
\keyword{taxonomy}
