% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MG_RastPipeline.R
\name{filt}
\alias{filt}
\title{set filter for function abundance}
\usage{
filt(b)
}
\arguments{
\item{b}{functions}
}
\description{
download functional annotations of many files, combine into a table
}
\examples{
netstat()
}
