% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MG_RastPipeline.R
\name{FT.cor}
\alias{FT.cor}
\title{function-correlation matrix}
\usage{
FT.cor(x, y)
}
\arguments{
\item{x}{taxon}

\item{y}{functional abundance}

\item{data}{data}
}
\description{
function-correlation matrix
}
\examples{
linmod()
}
