% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MG_RastPipeline.R
\name{linmod}
\alias{linmod}
\title{Linear model}
\usage{
linmod(x, y, data)
}
\arguments{
\item{x}{taxon}

\item{y}{covariate data}

\item{data}{data}
}
\description{
Linear model
}
\examples{
linmod()
}
