% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MG_RastPipeline.R
\name{load}
\alias{load}
\title{load}
\usage{
load(x, auth, ont)
}
\arguments{
\item{x}{list of mg-rast accession numbers}

\item{auth}{authentication code for mg-rast}

\item{ont}{level of ontology. acceptable: Subsystems, COG, KO, NOG}
}
\description{
submit function annotation query to mg-rast, get result url
}
\examples{
load()
}
\keyword{download}
\keyword{function}
\keyword{mg-RAST}
