% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MG_RastPipeline.R
\name{loadFunc}
\alias{loadFunc}
\title{load wrapper for many different samples}
\usage{
loadFunc(x, auth, ont, parallel)
}
\arguments{
\item{x}{list of mg-rast accession numbers}

\item{auth}{authentication code for mg-rast}

\item{ont}{level of ontology. acceptable: Subsystems, COG, KO, NOG}

\item{parallel}{logical; default is T}
}
\description{
load wrapper for many different samples
}
\examples{
loadFunc()
}
\keyword{download}
\keyword{function}
\keyword{mg-RAST}
