% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MG_RastPipeline.R
\name{MeanNetval}
\alias{MeanNetval}
\title{transform count values}
\usage{
MeanNetval(x)
}
\arguments{
\item{x}{taxa}
}
\description{
transform count values
}
\examples{
MeanNetval()
}
\keyword{igraph}
\keyword{metric}
\keyword{summaries}
