% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MG_RastPipeline.R
\name{mg.load}
\alias{mg.load}
\title{load}
\usage{
mg.load(x, auth, ont, level, E, length, id)
}
\arguments{
\item{x}{list of mg-rast accession numbers}

\item{auth}{authentication code for mg-rast}

\item{ont}{level of ontology. acceptable: Subsystems, COG, KO, NOG}

\item{level}{level of ontology. acceptable values: 1, 2, 3, 4.}

\item{E}{expected error value. default mg-rast is 5; (e^-5), larger numbers are more retrictive.}

\item{length}{minimun number of bp to match query string. mg-rast uses 15}

\item{id}{minimum percent match. mg-rast uses 60}
}
\description{
submit function annotation query to mg-rast, get result url
}
\examples{
load()
}
\keyword{download}
\keyword{function}
\keyword{mg-RAST}
