% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MG_RastPipeline.R
\name{plotNtwk2}
\alias{plotNtwk2}
\title{plot networks}
\usage{
plotNtwk2(list, name, num)
}
\arguments{
\item{x}{phyloseq object}

\item{cat}{categorical variable to subset phyloseq object}
}
\description{
make network explicitly defined number of edges
}
\examples{
plotNtwk()
}
\keyword{igraph}
\keyword{metric}
\keyword{summaries}
