% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MG_RastPipeline.R
\name{transform}
\alias{transform}
\title{transform count values}
\usage{
transform(x)
}
\arguments{
\item{x}{taxa}
}
\description{
transform count values
}
\examples{
transform()
}
\keyword{igraph}
\keyword{metric}
\keyword{summaries}
