% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MG_RastPipeline.R
\name{netStat}
\alias{netStat}
\title{construct table of network index values}
\usage{
netStat(list, groups2, name)
}
\arguments{
\item{groups2}{groups for stats model}

\item{x}{list of phyloseq objects}

\item{groups1}{list of groups to divide phylsoeq object by}
}
\description{
download functional annotations of many files, combine into a table
}
\examples{
netstat()
}
