% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MG_RastPipeline.R
\name{plotNtwk2}
\alias{plotNtwk2}
\title{construct table of connectivity values}
\usage{
plotNtwk2(list, name)
}
\arguments{
\item{x}{phyloseq object}

\item{cat}{categorical variable to subset phyloseq object}
}
\description{
download functional annotations of many files, combine into a table
}
\examples{
plotNtwk()
}
\keyword{igraph}
\keyword{metric}
\keyword{summaries}
